Reorganize provider classes into subdirectories
To reflect hierarchy
