+++ lib/carto/connector/providers/fdw/odbc/hive.rb
